+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/FileNameConfiguration.swift
@@ -17,3 +17,20 @@ struct FileNameConfiguration: SeverityBasedRuleConfiguration {
     @ConfigurationElement(key: "require_fully_qualified_names")
     private(set) var requireFullyQualifiedNames = false
 }
+
+// MARK: - For `excluded` option
+extension FileNameConfiguration {
+    func shouldExclude(filePath: String) -> Bool {
+        let fileName = filePath.bridge().lastPathComponent
+
+        // For backwards compatibility,
+        // `excluded` can have a fileName which is invalid as regex.(e.g. "NSString+Extension.swift")
+        if excluded.contains(fileName) {
+            return true
+        }
+
+        return excluded.compactMap {
+            try? Regex("^\($0)$").firstMatch(in: filePath)
+        }.isNotEmpty
+    }
+}