+++ CHANGELOG.md
@@ -31,6 +31,10 @@
   [JP Simard](https://github.com/jpsim)
   [Matt Pennig](https://github.com/pennig)
 
+* Add support for regex pattern in the `excluded` option of the `file_name` rule.  
+  [Ueeek](https://github.com/Ueeek)
+  [#6066](https://github.com/realm/SwiftLint/issues/6066)
+
 ### Bug Fixes
 
 * Improved error reporting when SwiftLint exits, because of an invalid configuration file